torch/pyproject.toml
[tool.poetry]
name = "hypertransformer"
version = "0.4.0"
description = "Pytorch implementation of (Continual) HyperTransformer."
authors = ["__z__ <imzqqq@hotmail.com>"]
license = "AGPL-3.0-or-later"
readme = "README.md"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Topic :: CS :: CV",
]
# package-mode = false

[[tool.poetry.source]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple/"
priority = "primary"
# priority = "supplemental"
[[tool.poetry.source]]
name = "aliyun"
url = "https://mirrors.aliyun.com/pypi/simple/"
priority = "supplemental"
[[tool.poetry.source]]
name = "nvidia"
url = "https://pypi.nvidia.com"
priority = "explicit"

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
pytest = "^9.0.2"
setuptools_rust = ">=1.11.1"
attrs = "^24.3.0"
numpy = "^2.1.1"
fastai = "^2.8.2"
scipy = "^1.16.1"
pandas = "^2.2.1"
ipywidgets = "^8.1.7"
seaborn = "^0.13.2"
requests = "^2.32.4"
matplotlib = "^3.10.3"
tqdm = "^4.67.1"
opencv-python = "^4.10.0.84"
pyyaml = "^6.0.2"
tensorboard = "^2.19.0"
torch-tb-profiler = "^0.4.3"
pydantic = "^2.7.4"
pillow = "^11.3.0"
scapy = "^2.6.1"
netron = "^8.4.6"
absl-py = "^2.1.0"
dm-tree = "^0.1.9"

torch = { version = "^2.7.1", optional = true }
torcheval = { version = "^0.0.7", optional = true }
# View model summaries in PyTorch!
torchinfo = { version = "^1.8.0", optional = true }
torchvision = { version = "^0.22.1", optional = true }
tensorrt = { version = "^10.12.0", optional = true }
onnx = { version = "^1.18.0", optional = true }
onnx-graphsurgeon = { version = "^0.5.8", optional = true }
onnxruntime-tools = { version = "^1.7.0", optional = true }
onnxruntime = { version = "^1.18.0", optional = true }

# `poetry install --extras torch`
[tool.poetry.extras]
torch = [
    "torch",
    "torcheval",
    "torchinfo",
    "torchvision",
    "tensorrt",
    "onnx",
    "onnx-graphsurgeon",
    "onnxconverter-common",
    "onnxruntime-tools",
    "onnxruntime",
]

# `poetry install --with tensorflow`
[tool.poetry.group.tensorflow.dependencies]
tensorflow = "^2.19.0"
tensorflow-datasets = "^4.9.9"
tf-slim = "^1.1.0"

# `poetry install --with llm`
[tool.poetry.group.llm.dependencies]
safetensors = "^0.5.0"
transformers = "^4.57.1"
triton = "^3.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
